Simplify freeze/translation process in a wholly new release
I was trying to preempt a future where we do more work in parallel by
allowing for continued work on master during a release, but this is
overkill at the moment.

We weren't specific enough about how the cycle of translations, fixes and
re-translation often works either, so we've fleshed that out.